requirements.txt
asgiref==3.2.7
Babel==2.8.0
Delorean==1.0.0
Django==3.0.6
djangorestframework==3.11.0
Faker==0.9.1
humanize==2.4.0
mixer==6.1.3
names==0.3.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.2
tzlocal==2.1
